Add user profile image crud

+++ app/controllers/api/v1/users.js
@@ -1,8 +1,15 @@
+const fs = require('fs');
 const {
   PrismaClient,
   PrismaClientKnownRequestError,
 } = require('@prisma/client');
-
+const path = require('path');
+const defaultImageUrl = path.join(
+  __dirname,
+  '../../../../public/images/default_image.png'
+);
+const imageKit = require('../../../../utils/imagekit');
+const { encryptPassword, checkPassword } = require('../../../../utils/auth');
 const prisma = new PrismaClient();
 
 const createUser = async (req, res) => {
@@ -14,15 +21,68 @@ const createUser = async (req, res) => {
         message: 'Bad Request: Request body is empty',
       });
     }
+    const {
+      name,
+      email,
+      password,
+      identity_type = null,
+      identity_number = null,
+      address = null,
+    } = req.body;
+    let identityNumber = identity_number ? parseInt(identity_number) : null;
 
-    const { profile, ...data } = req.body;
+    const user = await prisma.user.findFirst({
+      where: { email },
+    });
 
-    const user = await prisma.user.create({
+    if (user) {
+      return res.status(400).json({
+        status: 'error',
+        code: 400,
+        message: 'Email sudah terdaftar!',
+      });
+    }
+    let profile;
+    if (req.file) {
+      let stringFile = req.file.buffer.toString('base64');
+      const imageUpload = await imageKit.upload({
+        file: stringFile,
+        fileName: req.file.originalname,
+      });
+
+      profile = {
+        identity_type,
+        identity_number: identityNumber,
+        address,
+        image: imageUpload.url,
+        imageId: imageUpload.fileId,
+      };
+    } else {
+      const defaultImageUpload = await imageKit.upload({
+        file: fs.readFileSync(defaultImageUrl, 'base64'),
+        fileName: defaultImageUrl,
+      });
+
+      profile = {
+        identity_type,
+        identity_number: identityNumber,
+        address,
+        image: defaultImageUpload.url,
+        imageId: defaultImageUpload.fileId,
+      };
+    }
+
+    const createUser = await prisma.user.create({
       data: {
-        ...data,
+        name,
+        email,
+        password: await encryptPassword(password),
         profile: {
           create: profile,
         },
+      },
+      include: {
+        profile: true,
       },
     });
 
@@ -30,9 +90,10 @@ const createUser = async (req, res) => {
       status: 'success',
       code: 200,
       message: 'Data ditambahkan!',
-      data: user,
+      data: createUser,
     });
   } catch (error) {
+    console.error(error);
     return res.status(500).json({
       status: 'error',
       code: 500,
@@ -82,6 +143,7 @@ const getUsers = async (req, res) => {
 
 const getUserById = async (req, res) => {
   const id = parseInt(req.params.id);
+  const authenticatedUserId = req.user.id;
 
   try {
     if (isNaN(id)) {
@@ -109,6 +171,15 @@ const getUserById = async (req, res) => {
       });
     }
 
+    if (user.id !== authenticatedUserId) {
+      return res.status(401).json({
+        status: 'error',
+        code: 401,
+        message: 'Unauthorized: You do not have permission to access this user',
+      });
+    }
+
+    delete user.password;
     return res.status(200).json({
       status: 'success',
       code: 200,
@@ -126,6 +197,7 @@ const getUserById = async (req, res) => {
 
 const updateUser = async (req, res) => {
   const id = parseInt(req.params.id);
+  const authenticatedUserId = req.user.id;
   const { name, email, password } = req.body;
 
   try {
@@ -137,17 +209,31 @@ const updateUser = async (req, res) => {
       });
     }
 
+    if (id !== authenticatedUserId) {
+      return res.status(401).json({
+        status: 'error',
+        code: 401,
+        message: 'Unauthorized: You do not have permission to access this user',
+      });
+    }
+
+    const updateData = {
+      name,
+      email,
+    };
+
+    if (password) {
+      updateData.password = await encryptPassword(password);
+    }
+
     const updateUser = await prisma.user.update({
       where: {
         id,
       },
-      data: {
-        name,
-        email,
-        password,
-      },
+      data: updateData,
     });
 
+    delete updateUser.password;
     return res.status(200).json({
       status: 'success',
       code: 200,
@@ -176,6 +262,7 @@ const updateUser = async (req, res) => {
 
 const deleteUser = async (req, res) => {
   const id = parseInt(req.params.id);
+  const authenticatedUserId = req.user.id;
   try {
     if (isNaN(id)) {
       return res.status(400).json({
@@ -184,12 +271,21 @@ const deleteUser = async (req, res) => {
         message: 'Bad Request : ID must be a number',
       });
     }
+
+    if (id !== authenticatedUserId) {
+      return res.status(401).json({
+        status: 'error',
+        code: 401,
+        message: 'Unauthorized: You do not have permission to access this user',
+      });
+    }
+
     const deletedUser = await prisma.user.delete({
       where: {
         id,
       },
     });
-
+    delete deletedUser.password;
     return res.status(200).json({
       status: 'success',
       code: 200,
@@ -216,10 +312,122 @@ const deleteUser = async (req, res) => {
   }
 };
 
+const updateProfileAndImage = async (req, res) => {
+  try {
+    const userId = req.user.id;
+    const { identity_type, identity_number, address } = req.body;
+    let identityNumber = identity_number
+      ? parseInt(identity_number)
+      : undefined;
+    let imageUpload, imageId;
+    if (req.file) {
+      const stringFile = req.file.buffer.toString('base64');
+      imageUpload = await imageKit.upload({
+        file: stringFile,
+        fileName: req.file.originalname,
+      });
+      imageId = imageUpload && imageUpload.fileId;
+    }
+
+    const currentProfile = await prisma.profile.findUnique({
+      where: {
+        userId,
+      },
+    });
+
+    const updatedProfile = await prisma.profile.update({
+      where: {
+        userId,
+      },
+      data: {
+        identity_type,
+        identity_number: identityNumber,
+        address,
+        image: imageUpload && imageUpload.url,
+        imageId: imageId || null,
+      },
+    });
+
+    if (currentProfile.imageId && imageUpload) {
+      await imageKit.deleteFile(currentProfile.imageId);
+    }
+
+    if (!updatedProfile) {
+      return res.status(401).json({
+        status: 'error',
+        code: 401,
+        message:
+          'Unauthorized: You do not have permission to update this profile',
+      });
+    }
+
+    return res.status(200).json({
+      status: 'success',
+      code: 200,
+      message: 'Profile information and image updated successfully',
+      data: updatedProfile,
+    });
+  } catch (error) {
+    console.error(error);
+    return res.status(500).json({
+      status: 'error',
+      code: 500,
+      message: 'Internal server error',
+    });
+  }
+};
+
+const deleteProfileImage = async (req, res) => {
+  try {
+    const userId = req.user.id;
+
+    const userProfile = await prisma.profile.findUnique({
+      where: {
+        userId,
+      },
+    });
+
+    if (userProfile && userProfile.image) {
+      await imageKit.deleteFile(userProfile.imageId);
+    }
+
+    const defaultImageUpload = await imageKit.upload({
+      file: fs.readFileSync(defaultImageUrl),
+      fileName: defaultImageUrl,
+    });
+
+    const updatedProfile = await prisma.profile.update({
+      where: {
+        userId,
+      },
+      data: {
+        image: defaultImageUpload.url,
+        imageId: defaultImageUpload.fileId,
+      },
+    });
+
+    return res.status(200).json({
+      status: 'success',
+      code: 200,
+      message: 'Profile image deleted, set to default image successfully',
+      data: updatedProfile,
+    });
+  } catch (error) {
+    console.error(error);
+    return res.status(500).json({
+      status: 'error',
+      code: 500,
+      message: 'Internal server error',
+    });
+  }
+};
+
 module.exports = {
   createUser,
   getUsers,
   getUserById,
   updateUser,
   deleteUser,
+  updateProfileAndImage,
+  deleteProfileImage,
 };
